pyproject.toml
[project]
name = "weather-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.11",
    "requests>=2.32.5",
]
